Level_0 Update Part 1
I added decals, vegetation, assets. I blocked the hallway on the observatory. Inside the maintenance room I re-arrange it in a way the Oxygen station serves as foreshadowing and the flashlight on it for the player to grave it. Also the elevator have a change having the button properly appealing. I downgraded more textures to 1080. At least I hide the hands for the Pc version.